requirements.txt
pandas
numpy
scikit-learn
streamlit
google-generativeai
python-dotenv
openpyxl